Initial changes to move the server architecture to a layered approach.
