+++ src/BlazorBoilerplate.Server/Controllers/ApiLogController.cs
@@ -1,10 +1,10 @@
 ﻿using BlazorBoilerplate.Server.Middleware.Wrappers;
-using BlazorBoilerplate.Server.Services;
 using BlazorBoilerplate.Shared.AuthorizationDefinitions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Threading.Tasks;
+using BlazorBoilerplate.Server.Managers;
 
 namespace BlazorBoilerplate.Server.Controllers
 {
@@ -12,27 +12,23 @@ namespace BlazorBoilerplate.Server.Controllers
     [ApiController]
     public class ApiLogController : ControllerBase
     {
-        private readonly IApiLogService _apiLogService;
+        private readonly IApiLogManager _apiLogManager;
 
-        public ApiLogController(IApiLogService apiLogService)
+        public ApiLogController(IApiLogManager apiLogManager)
         {
-            _apiLogService = apiLogService;
+            _apiLogManager = apiLogManager;
         }
 
         // GET: api/ApiLog
         [HttpGet]
         [AllowAnonymous]
         public async Task<ApiResponse> Get()
-        {
-            return await _apiLogService.Get();
-        }
+        =>  await _apiLogManager.Get();
 
         // GET: api/ApiLog/ApplicationUserId
         [HttpGet("[action]")]
         [Authorize(Policy = Policies.IsAdmin)]
         public async Task<ApiResponse> GetByApplicationUserId(string userId)
-        {
-            return await _apiLogService.GetByApplictionUserId(new Guid(userId));
-        }
+        =>  await _apiLogManager.GetByApplicationUserId(new Guid(userId));
     }
 }
