+++ src/BlazorBoilerplate.Server/Managers/IApiLogManager.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Threading.Tasks;
+using BlazorBoilerplate.Server.Middleware.Wrappers;
+using BlazorBoilerplate.Server.Models;
+
+namespace BlazorBoilerplate.Server.Managers
+{
+    public interface IApiLogManager
+    {
+        Task Log(ApiLogItem apiLogItem);
+        Task<ApiResponse> Get();
+        Task<ApiResponse> GetByApplicationUserId(Guid applicationUserId);
+    }
+}
