+++ src/BlazorBoilerplate.Server/Managers/IEmailManager.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+using BlazorBoilerplate.Server.Middleware.Wrappers;
+using BlazorBoilerplate.Shared;
+using BlazorBoilerplate.Shared.Dto;
+
+namespace BlazorBoilerplate.Server.Managers
+{
+    public interface IEmailManager
+    {
+        Task<ApiResponse> Send(EmailDto parameters);
+        Task<ApiResponse> Receive();
+        Task<ApiResponse> SendEmailAsync(EmailMessageDto emailMessage);
+        List<EmailMessageDto> ReceiveEmail(int maxCount = 10);
+        Task<ApiResponse> ReceiveMailImapAsync();
+        Task<ApiResponse> ReceiveMailPopAsync(int min = 0, int max = 0);
+        void Send(EmailMessageDto emailMessage);
+    }
+}
