+++ src/BlazorBoilerplate.Server/Managers/IExternalAuthManager.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
+
+namespace BlazorBoilerplate.Server.Managers
+{
+    public interface IExternalAuthManager
+    {
+        Task<IActionResult> Challenge(string provider);
+        Task<IActionResult> ExternalSignIn();
+    }
+}
