+++ src/BlazorBoilerplate.Server/Managers/ITodoManager.cs
@@ -1,0 +1,15 @@
+﻿using System.Threading.Tasks;
+using BlazorBoilerplate.Server.Middleware.Wrappers;
+using BlazorBoilerplate.Shared.Dto;
+
+namespace BlazorBoilerplate.Server.Managers
+{
+    public interface ITodoManager
+    {
+        Task<ApiResponse> Get();
+        Task<ApiResponse> Get(long id);
+        Task<ApiResponse> Create(TodoDto todo);
+        Task<ApiResponse> Update(TodoDto todo);
+        Task<ApiResponse> Delete(long id);
+    }
+}
